Fix memory leak in pollGLFW
In particular, it turns out writeTVar doesn't evaluate its argument when it
actually commits the thunk to memory. This causes thunks to build up in the
resulting value if they are never read (such as for scroll input). Hence, we
need to fully evaluate the entire input state once we write it to the tvar.